[Feature] Docker build (#1)
* Add Dockerfile pre-setup for burrito

Burrito requires Zig as well. So ASDF is used to install Zig, Elixir and Erlang before adding the required changes to mix.exs

* Add full burrito support

* Working with tests

Had to change the files around to allow for normal run, docker run and test run. had to add a runtime.exs to be able to read the MIX_ENV properly.

* Add documentation for building the app

* Update disclaimer for building